WL#15005 ndbcrunch benchmark
Create new test suite for benchmarking a MySQL Cluster with all nodes
colocated on a single machine. The sysbench program is used for
generating load.

The configuration defines a cluster with a configurable number of MySQL
Servers, two data nodes and one mgm node. The suite is primarily
intended to be used for benchmarking the MySQL Server + ndbcluster
plugin performance. The suite consist of serveral tests that each will
run a diffferent sysbench benchmark. Environment variables are used to
control parameters of the benchmark.

The tests run sysbench in various modes against the ndbcrunch cluster.
Output will be saved as report_<testname>.log.

Parameters for sizing can be provided as follows:
 CRUNCH_ROWS
   - number of rows(default 10M)
 CRUNCH_TIME
   - time to run(default 30 seconds)
 CRUNCH_EVENTS
   - number of events(default 0 -> runs until  time elapsed)
 CRUNCH_THREADS
   - use specific thread number(default 0 -> geometric thread series)
 CRUNCH_AUTOINC
  - use auto increment column(default 0 -> no auto inc)
 CRUNCH_ENGINE
   - select storage engine(default ndbcluster)

For example like this:
 $ CRUNCH_TIME=10 CRUNCH_ROWS=100000 mtr --suite=ndbcrunch

Other settings are controlled by the tests.

Change-Id: I2acfa3b891b3aa434fcd23edb2a0f47e8ae5c16d