HOPSWORKS-2525: Step 1
Added a new socket to each receive thread. This socket is used to wakeup
the receive thread when communication from another thread wants to make
use of the receive thread.

This feature is important when the receive thread is used for other activities
than the receive handling. In this case the other thread needs the receive
thread to immediately react. For other threads this wakeup happens through
a futex_wake on Linux and a condition signal on other platforms.

However receive thread sleeps on either epoll_wait or on poll. Thus to
only mechanism to wake those threads is by sending something to a socket
that the receive thread listens to. This is where the extra socket comes
into play. To wake the receive thread it is enough to send 1 byte to the
receive thread and it will immediately wakeup.

To handle this conditional wakeup we added a boolean on the thread object
indicating if it is a receive thread or not, we also added a reference
to the TransporterReceiveHandle of the receive thread.

This patch enables all sorts of experimentation with setting up threads
in new manners with the receive thread as an active participant in both
receive and other activities.