HOPSWORKS-2525: Step 2 and Step 3
Step 2:
When using ThreadConfig and not creating any TC threads we will instead map
the TC threads to the receiver threads.

Step 3:
When TCSEIZEREQ arrives we will check which node sent the message and assign
the DBTC instance that is colocated with the receive thread instance handling
this node. This means that when receiving TCKEYREQ, SCAN_TABREQ, these signals
will be sent locally to the same thread. This cuts some latency away and
could potentially be a performance benefit.