HOPSWORKS-2525: Step 7
Changed some parts of the automatic thread configuration.
Now that we can handle TC and receive in the recv thread
it is possible to make a bit more efficient of CPU resources
in smaller configurations.

However using tc threads is still used in larger configs
since it is still more efficient.