HOPSWORKS-2525: Step 8
This patch introduces one more variant of how to configure threads
in RonDB data nodes. Previously the only configuration that didn't
have specific LDM threads was a configuration with a single receive
thread and a configuration with a single receive thread and a
single main thread.

In this patch we enable a configuration with a large number of
receive threads without any LDM threads. In this configuration
the idea is that the receive threads will be able to do all work
from start to beginning. Thus executing without a thread pipeline.
The only need in traffic execution to not do everything in the
local receive thread is handling of non-committed READs and any
WRITE queries. These still can only be handled by the LQH that
owns the data.

This configuration cannot be combined with TC threads, Query
threads and Recover threads. Thus in this configuration we only
have receive threads and possibly main thread(s).

In this configuration each receive thread has 1 LQH worker,
one Query thread worker, 1 DBTC worker. This means that any
Committed Read queries can be served fully in the receive
thread.

Having send threads or not is still optional in this configuration.

In this configuration we don't activate any load distribution
mechanisms to pick the right query thread. We always pick the
local query thread worker.

We have made stronger division in this patch with the use of
globalData.ndbMtLqhThreads vs globalData.ndbMtLqhWorkers
and similarly for ndbMtQueryThreads/ndbMtQueryWorkers.
Likewise we previously did the same thing for ndbMtTcThreads/
ndbMtTcWorkers.

There is no such distinction for ndbMtMainThreads and ndbMtReceiveThreads,
there are no special variables for workers for these thread types
and similarly not for send threads.