HOPSWORKS-2524
Add support for Longvarchar data types in ClusterJ.
This is necessary to handle primary keys that are defined as
VARCHAR(x) where max size of this column is larger than 255
bytes. Thus even VARCHAR(64) can be a problem with character
sets that are 4 byte per character.
Added tests also for Varchar as primary key which was previously
missing although it was supported.