Fix test_mgm
When a new configuration was created we removed all LogHandlers
currently used and replaced them with the new ones as decided
by the new configuration. However in doing so we created a new
FileLogHandler object, but when removing the old objects this
object was not destroyed, this led to a leak of an open file
descriptor eventually leading to running out of open file
descriptors.

Fixed by calling the destructor of FileLogHandler as part of
destructor of BufferedFileHandler object which is the one
placed in the list of LogHandlers on g_eventLogger.