HOPSWORKS-2565
ndb_import will use values from CSV files to set values of auto increment
values. It will however not ensure that the auto increment values set
in RonDB will be correct after the ndb_import process.

To solve this we introduce a new option to ndb_import called
--use-auto-increment, this will ignore the settings from the CSV file
and instead set an autoincrement value, this will ensure that INSERTs
to the table will be correct and working after importing data using
ndb_import.

If this option isn't set it is necessary to set a new starting value
for the auto increment. This can be done e.g. by the following
transaction:

BEGIN;
INSERT INTO table (column list) (auto_increment_value, values)
ROLLBACK;

This will set the new auto increment value in the table.