HOPSWORKS-2570
seizeLogpage fails in ptrCheckGuard in a properly allocated
log pages. It fails since we assume that all the log pages
are consecutive while allowing them to be allocated in chunks.

The simple solution is to ensure that they are actually
allocated consecutively which is what happens when we
require alloc_pages in ndbd_malloc.cpp to return all
pages in 1 chunk from the allocChunks method.