HOPSWORKS-2572
The current distribution of primary replicas isn't optimal for
the new fragmentation variants. With e.g. 8 fragments per
table and 2 nodes we will find that the same LDM thread gets
two fragments to act as primary for in a 4-LDM setup whereas
the 2 LDM threads gets no primary replicas to handle.

This is handled by a better setup at creation of the table.
However to also address handling of Not Active nodes we
need to also redistribute the fragments at various events.

The redistribution is only allowed if all nodes have upgraded
to 21.10 in the cluster. Older versions of RonDB will not
redistribute and we need to ensure that all data nodes use
the same primary replicas. If not we would cause a multitude
of constant deadlocks.