HOPSWORKS-2591
releaseTcCon uses a block variable as input parameter.
However this input parameter is changed by the call to
abortTransFromTrigger which leads to release of the wrong
record and later that can lead to a crash.

It only happens after a successful allocation of a TC connect
record followed by an unsuccessful allocation of an API
connect record. Thus will almost never happen in practice.
But it does happen in the test case ndb_transaction_memory_shortage.