HOPSWORKS-2603
In a single node setup all restarts are system restarts.
In this case if the node fails in the last few seconds of
the restart this leads to a node in the state not
recoverable on its own. Obviously a node in a single node
setup cannot rely on any other node for recovery so this
isn't correct.

The problem with nodes not being able to recover on its own
is related to node restart as explained in a large comment
in Backup.cpp. Thus System restarts should never set the
state to NODE_NOT_RECOVERABLE_ON_ITS_OWN