HOPSWORKS-2573: Step 2
Added more comments on the implementation of executing locked reads
in the query threads.

Removed the use of insertLockOwnersList and takeOutOwnersList from
production builds, keep it in debug and error insert builds.

Ensure that access to primary keys in DBACC use the ACC fragment
mutex also in LDM thread.

Added mutex protecting the access of the transaction hash in both
LDM and query threads. We focus on the use with query threads, thus
we always lock this mutex although it isn't necessary if only using
LDM threads.