HOPSWORKS-2573: Step 4
Added new variables tracking which is the current LDM blocks.
m_curr_lqh tracks which LQH block instance is the one we are
currently working on. Similarly m_curr_acc in DBACC and
m_curr_tup in DBTUP.

These are mainly used to access the correct mutexes, the
correct hash table and using the correct pool to get the
operation records.