HOPSWORKS-2610
For RonDB to interact with systemd in a good way we need to
enable graceful shutdown even if it is started using
kill -TERM ndbmtd.

The method in NDB to handle shutdown isn't working very well
since it is hard to know which processes to stop to really
stop ndbmtd. However most use cases can use
killall -TERM ndbmtd
to kill all ndbmtd processes in the VM.
This patch ensures that the above command leads to a graceful
shutdown that completes within 30 seconds.