HOPSWORKS-2573
DBT2 benchmark was heavily impacted by the use of the
transaction hash table. DBT2 has a transaction with
tens of thousands of exclusive reads per transaction.

To ensure that this doesn't cause very long searches of
linked lists we increase the size of the hash table to
128k from 4k. It still only uses 0.5MB per LDM and
query thread. Actually eventually we should be able to
remove this from the Query thread since this hash table
is only used in LDM threads.

This causes almost a 20% improvement of DBT2 performance.