HOPSWORKS-2493
A major bottleneck is the index statistics mutex.
This is acquired 3 times per index lookup to gather index statistics.
This becomes a bottleneck when Sysbench OLTP RW reaches around 10000
TPS with around 100 threads. Thus a severe limitation on scalability
for the MySQL Server using RonDB.

To handle this we ensure that the hot path through the code doesn't
need to acquire the global mutex at all. This is solved by using the
NDB_SHARE mutex a bit more and making the ref_count variable an
atomic variable.

Also needed to handle some global statistics variables. Fixed by adding
them on local object and every now and then transferring to the global
bject.