HOPSWORKS-2652
When an update causes a delete of a row in an ordered index
we will move any scan pointers that currently points to the
deleted row. We move this row using scanNext and after this
we call relinkScan to move the position such that scan
continues from the correct position.

In this we used our scan instance instead of using the
scan instance of the scan record we are moving.