HOPSWORKS-2525
Step 1:
Added a new socket to each receive thread. This socket is used to wakeup
the receive thread when communication from another thread wants to make
use of the receive thread.

This feature is important when the receive thread is used for other activities
than the receive handling. In this case the other thread needs the receive
thread to immediately react. For other threads this wakeup happens through
a futex_wake on Linux and a condition signal on other platforms.

However receive thread sleeps on either epoll_wait or on poll. Thus the
only mechanism to wake those threads is by sending something to a socket
that the receive thread listens to. This is where the extra socket comes
into play. To wake the receive thread it is enough to send 1 byte to the
receive thread and it will immediately wakeup.

To handle this conditional wakeup we added a boolean on the thread object
indicating if it is a receive thread or not, we also added a reference
to the TransporterReceiveHandle of the receive thread.

These changes enables all sorts of experimentation with setting up threads
in new manners with the receive thread as an active participant in both
receive and other activities.

Step 2:
When using ThreadConfig and not creating any TC threads we will instead map
the TC threads to the receiver threads.

Step 3:
When TCSEIZEREQ arrives we will check which node sent the message and assign
the DBTC instance that is colocated with the receive thread instance handling
this node. This means that when receiving TCKEYREQ, SCAN_TABREQ, these signals
will be sent locally to the same thread. This cuts some latency away and
could potentially be a performance benefit.

Step 4:
Removed the requirement that receive threads had the nosend flag set to 1.
Updated the receive thread's main loop to reflect that it can also act as
a block thread in addition to performing receive activities.

The receive thread does an extra flush after receiving data on transporters
to ensure that execution of its own signals doesn't cause receive thread
to slow down flushing signals to other threads. This included ensuring
that send buffer pool is filled before starting to execute signals.

Simplified handling of alert_send_thread in that all send threads are
woken up, also the one we will assist. This decreases the number of
acquisitions of the send thread mutex and greatly simplifies the code.

do_send is called in the same manner as in block threads when no signals
was executed.

Clarified that code on handling load indicators are only required for
LDM and Query threads.

Ensured that sendpacked is called in more situations to assist in NDBFS
communication.

Step 5:
Changed some parts of the automatic thread configuration.
Now that we can handle TC and receive in the recv thread
it is possible to make a bit more efficient of CPU resources
in smaller configurations.

However using tc threads is still used in larger configs
since it is still more efficient.

Step 6:
This patch introduces one more variant of how to configure threads
in RonDB data nodes. Previously the only configuration that didn't
have specific LDM threads was a configuration with a single receive
thread and a configuration with a single receive thread and a
single main thread.

In this patch we enable a configuration with a large number of
receive threads without any LDM threads. In this configuration
the idea is that the receive threads will be able to do all work
from start to beginning. Thus executing without a thread pipeline.
The only need in traffic execution to not do everything in the
local receive thread is handling of non-committed READs and any
WRITE queries. These still can only be handled by the LQH that
owns the data.

This configuration cannot be combined with TC threads, Query
threads and Recover threads. Thus in this configuration we only
have receive threads and possibly main thread(s).

In this configuration each receive thread has 1 LQH worker,
one Query thread worker, 1 DBTC worker. This means that any
Committed Read queries can be served fully in the receive
thread.

Having send threads or not is still optional in this configuration.

In this configuration we don't activate any load distribution
mechanisms to pick the right query thread. We always pick the
local query thread worker.

We have made stronger division in this patch with the use of
globalData.ndbMtLqhThreads vs globalData.ndbMtLqhWorkers
and similarly for ndbMtQueryThreads/ndbMtQueryWorkers.
Likewise we previously did the same thing for ndbMtTcThreads/
ndbMtTcWorkers.

There is no such distinction for ndbMtMainThreads and ndbMtReceiveThreads,
there are no special variables for workers for these thread types
and similarly not for send threads.

Step 7:
Make it possible to set nosend=1 also on Query threads.

Step 8:
Use only LDM threads with 4 CPUs, no specific gain with
only 2 CPUs to use LDM threads.

Step 9:
Previously performReceive first read from all transporters and then looped
over all transporters to unpack the read data. This means that we sweep
through the data twice, seems better to unpack data immediately after
receiving the data. In the NDB API this even means that the signal execution
happens when data is already in CPU caches. So could potentially provide
even bigger benefits for the NDB API performance.

Step 10:
Reorganised code in performReceive a bit.