HOPSWORKS-2573
Make query threads be able to also handle Locked Reads including
take over operations using LQHKEYREQ.

This also includes enlarging hash tables and reserved operation
records available to both LDM and Query threads. This is in
itself a significant performance boost.

There were a number of choices available how to handle the locked
list. To avoid changes in the row data structures we opted for
using operation records from the LDM thread also in the query
thread. This means also that COMMIT processing can be handled
by the LDM thread since the operation record already belongs to
the LDM thread.

Another major change in this job was to completely remove support
for ACC full table scans. This simplifies DBACC significantly.
We haven't taken advantage of this yet, there is still a lot of
dead code in DBACC no longer used. The fact is that TUP scans
do the job good enough so there is no reason to carry around the
complexity of ACC scans.

The change requires some upgrade code to ensure that we don't
attempt to use locked reads from query threads in nodes that
do not support this.

Step 1:
Add comments in code about changes

Step 2:
Added more comments on the implementation of executing locked reads
in the query threads.

Removed the use of insertLockOwnersList and takeOutOwnersList from
production builds, keep it in debug and error insert builds.

Ensure that access to primary keys in DBACC use the ACC fragment
mutex also in LDM thread.

Added mutex protecting the access of the transaction hash in both
LDM and query threads. We focus on the use with query threads, thus
we always lock this mutex although it isn't necessary if only using
LDM threads.

Step 3:
Added allocation of operation records from LDM owning thread
for locked reads.

Step 4:
Added new variables tracking which is the current LDM blocks.
m_curr_lqh tracks which LQH block instance is the one we are
currently working on. Similarly m_curr_acc in DBACC and
m_curr_tup in DBTUP.

These are mainly used to access the correct mutexes, the
correct hash table and using the correct pool to get the
operation records.

Step 5:
ACCKEY_ORD required also to be protected by the ACC fragment mutex.
Also all calls to the startNext, startNew, release_lockowner to
maintain the Row Lock Queues were required to be protected. Added
ndbabort in all places of the code where ACC scans were the only
use case.

Step 6:
Fixed a problem that code in insertLockOwnersList had code
that was required also in non-debug mode which had to be moved
outside of the method.

Added lots of comments about concurrency issues to consider in DBACC.
Added lots of comments in preparation for changes in DBTC.

Step 7:
Added a bit more debug support of concurrency handling in DBACC.

Step 8:
Cannot read the variable m_op_bits before acquiring the mutex when the
variable is in a Lock Queue list. Committed reads are not in any
Lock Queue list and are thus not in need of protection.

Step 9:
Added lots of debug around transaction hash mutex and other mutex issues

Step 10:
While operation is in transaction hash in DBLQH, the variables
prevHashRec and nextHashRec needs mutex protection. Thus when
removed from list we cannot read those before we acquired the
mutex.

Step 11:
Disable query thread usage debugging

Step 12:
Use Query threads also for Read Take over operations.
Used a model to analyse how Take Over operations are
handled by the NDB API. Turned out that internal scans
are not using this model, internal scans act already
on receiving TRANSID_AI to perform take over. Thus we
needed some special handling of this.

Fixed an issue with selection of when to use optimised
record seize in execLQHKEYREQ.

Wrote a lengthy description of handling take over operations
with query threads.

Fixed some missing initialisation of scanTakeOver variable
in signal UTIL_EXECUTE_REQ.

Fixed signal names for COPY_DATA_IMPL* signals.

Increased size of take over hash.

Step 13:
DBT2 benchmark was heavily impacted by the use of the
transaction hash table. DBT2 has a transaction with
tens of thousands of exclusive reads per transaction.

To ensure that this doesn't cause very long searches of
linked lists we increase the size of the hash table to
128k from 4k. It still only uses 0.5MB per LDM and
query thread. Actually eventually we should be able to
remove this from the Query thread since this hash table
is only used in LDM threads.

This causes almost a 20% improvement of DBT2 performance.

Step 14:
Increase number of reserved operation records to decrease overhead
of the allocation of operation records.

Step 15:
Introduce a new list shared between LDM threads and Query threads
used for locked reads. This ensures that 40% of the reserved
operation records are only to be used by LDM thread and the
remaining 60% is to be shared with query threads.

Step 16:
Released operation records into query thread when using
new shared list instead of into current LQH.

Step 17:
Debug of seize connect in LQH