Schema Memory
Step 1:
Added a new class RWPool64
This class uses a 64-bit i-value. Apart from this it behaves exactly as
the RWPool. Thus the RWPool64 is not limited to the first 16 GB of the
global memory space.

To support this class a new class Record64 was created.
Also a new class Ptr64 and ConstPtr64 was created that
uses 64-bit i-values.

Also a new class Intrusive64List was created that creates a set of
templates for Single-linked list, Double-linked list, Counters that
all use 64-bit i-values as reference between records. The main
naming difference is that 64 is introduced into the naming in a
number of places.

Finally the new class RWPool64 was used to create the Log_waiter
class. This class was used as an example as it has no i-values
stored. Most classes store i-values somewhere and thus need
also require changes of those references.

The Undofile class has so few entries that it isn't necessary to
turn it into the RWPool64 class. This completes the work on LGMAN
in this change set.

Step 2:
Handle ReplicationMemory

Added handling of variable size of ReplicationMemory used in Suma.cpp
There are two use cases for ReplicationMemory. The buffers where GCIs
are stored in preparation for sending to the APIs that have subscribed
to events on tables. These buffers can be quite large since they need
to store events for up to a few seconds of writes.

The second use is for buffers used in signalling from DBTUP in LDM
threads to SUMA. These are short lived and should not require so
much memory.

Both of those structures use SharedGlobalMemory, but the first one
is limited by the number of chunks of pages that it can store.

Changed to enable storing up to 16 GBytes of chunks. However these
are allocated from SharedGlobalMemory in the same fashion as
QueryMemory, thus limiting them to use at most 90% of the
available SharedGlobalMemory. Also implemented returning chunks of
pages when no longer needed. Actual allocation and deallocation of
memory always happens in chunks of 1 MByte at a time to ensure that
we don't overload the memory allocation mutex. This is the case for
both of those structures.

Also added some support for SchemaMemory and ReplicationMemory.

Step 3:
Subscription moved to SchemaMemory using TransientPool
SubOpRecord moved to SchemaMemory using TransientPool
Table record in Suma moved to SchemaMemory using TransientPool
Subscriber record moved to SchemaMemory using TransientPool
Moved DataBuffer in SUMA to SchemaMemory using TransientPool

Fixed problems relating to multi-page buffers when sending between
DBTUP and SUMA.

Moved FragmentRecord in PGMAN to SchemaMemory using RWPool64,
increased num tables to max.
Moved DictObject to SchemaMemory using TransientPool
TriggerRecord in DBDICT moved to SchemaMemory using TransientPool
Moved TableRecord in DBDICT to SchemaMemory using TransientPool

The TableRecord had a self reference in using TableRecord_list::Head
in the record. This is an object which is well defined using IntrusiveList.hpp.
Previously this was handled by declaring the typedef of the ArrayPool of
TableRecord's before the actual definition of the TableRecord. This is not
possible using TransientPool since this uses the size of the TableRecord as
part of the definition of the template code for the TransientPool.

To handle this a somewhat ugly construct was used that declares a
IndexRecord_list::Head using an ArrayPool and then using this in
the construction of a LocalTableRecord_list object.

Moved AttributeRecord to SchemaMemory using TransientPool
Increase hash size of TcFKData

Moved TcIndexData in DBTC to SchemaMemory using TransientPool
Replaced seizeId with hash table of appropriate size.

Moved TcDefinedTriggerData to SchemaMemory using TransientPool
Replaced use of seizeId with Hash table

Moved ACC fragment record to RWPool64 from old ptrCheckGuard model
The RWPool64 requires no special handling of Query threads since they
are using the global memory root which is shared by all threads.

Moved TUP fragment record to SchemaMemory using RWPool64

DBLQH fragment record moved to SchemaMemory using RWPool64

Moved TUX fragment record to SchemaMemory using RWPool64

Moved TupTriggerData to SchemaMemory by using RWPool64

Moved ReplicaRecord to SchemaMemory by using RWPool64
Moved Fragmentstore to SchemaMemory by using RWPool64

Increased some hash sizes and allowable parallelism in UNDO log execution

Step 4:
Implement support for generic malloc calls for objects up to 1 MByte in
size. Also implement support for malloc of extremely short lived memory
areas mainly intended for message buffers. This new memory is still
allocated from the global memory pool and the requester must provide
the memory area to allocate from. This will enable generic malloc calls
for a specific memory area such as TransactionMemory, SchemaMemory,
ReplicationMemory, DataMemory and so forth.

A lot of functional unit tests as well as performance unit tests was
added to the generic malloc functions.

This step prepares to move fragment record and replica record in to
RWPool64.

It also prepares for a lot of new possibilities.

Various other things
--------------------
A unit test of the generic malloc calls, both for generic malloc calls and
for the short lived memory parts are implemented and this can also be used
to perform microbenchmarks of the malloc parts.

Removed looking for ndbd in MTR tests. Removed ndbd from builds.

Added test using 8160 partitions in ndb_basic (2rpl, 3rpl and 4 rpl tests).
Various test results updated

Fixed bug with MAX_WORDS_META_FILE handling:
 - MAX_WORDS_META_FILE increased to 65536 words
 - Ensured that we can't create tables that are bigger than this
 - Ensured that we discover errors in GET_TABINFO_REQ in Dbdict due
   space restrictions.

Fixed a number of printouts of signals and their names.
Added nodeFragCount to LQHFRAGREQ to enable allocating an array of proper
size in table object.

Added table id and fragment id to RESTORE_LCP_CONF.
Removed i-values of fragments in TUXFRAGREQ, now set in a different way.

Added error variable to SimpleProperties object to enable checks after
many operations.

Added SCHEMA_MEMORY and REPLICATION_MEMORY to ndbinfo table.

Removed mappings from fragid to fragPtrI in TUX, ACC and TUP, now reusing
map array in LQH for all of those. Added a number of functions to handle
this.

Ensured all variables in table objects in LQH, TUP, ACC and TUX are properly
initialised at startup and after drop.

Decreased MAX_PARALLEL_SCANS_PER_FRAG to 0 since we removed using ACC for
table scans. Actually this makes most code in DBACC deprecated. No cleanup
of this yet.

Reimplemented Rope class by using new malloc infrastructure.
Removed StringMemory, now using SchemaMemory for this.

A table can now have 8160 partitions even with a single LDM thread.
Removed implementation of memory allocation in DbtupTabDesMan.cpp,
replaced using new malloc functionality.

Minor optimisations removing recalculating pointers to fragment records
in traffic code parts.

Extended a number of hash sizes to avoid long linked lists in those cases.

No more use of WOPool, thus removed this from code base.