HOPSWORKS-2572
When we reorganise the primary we must ensure that the new primary is known by the
starting node. This also must work with old versions starting up. If the
preferredPrimary is already started it isn't an issue since that node will
always be selected as primary if it is alive. However if that node is currently
starting or isn't started we need to ensure that the nodes doesn't have a
differing opinion on who is the primary.

The starting node gets the distribution as part of the master node sending
the distribution information to the starting node. Thus we ensure that the
first node in the replica list will always be the current primary.

Fixed issues with calc_primary_replicas and improved its documentation.

More fixes of primary node distribution