HOPSWORKS-2749
A long description of how to handle distribution keys added.
Fixed a bug in that we missed updating the stored distribution keys
in node 1 in some 3 replica setups.
Avoid aborting transactions that see a newer distribution key in
signal compared to the stored, this is ok. What isn't ok is an
older distribution key received.