HOPSWORKS-2572
The NDB API will select the node for transaction coordinator in the following
priority order:
1) Select a node from the node group of hinted fragment if hint used
2) Select a node in the same location domatin if location domains are used
3) Select a node with the best proximity (thus favoring nodes using shared
   memory or nodes on the same hostname)
4) Select the primary node as the NDB API thinks (the NDB API doesn't have
   the latest information about where the primary is

Added upgrade code for HOPSWORKS-2572
Added upgrade handling of RonDB 21.10.2
Fixed an issue with DROP_FRAGREQ in LDM threads
where we released fragment record in LQH too early

When a node started it calculated the new primary node to be used
in the COMMIT_STORED signal of UPDATE_FRAG_STATEREQ. However the
primary node was in the STORED signal instead and immediately
set to 0 after that which meant that the calculated primary
replica after a failure didn't work.

Also ensured that we set the correct primary node even when
just one node is alive.

Fix such that fragmented signals to other nodes are not sent
to query threads in other nodes.