HOPSWORKS-2573
Fixed issues when a query thread executes a locked read and the transaction
is aborted while the LQHKEYREQ is still outstanding. Required some careful
changes of the signal orders.

The introduction of locked reads for query threads required a more elaborate handling
of ABORT messages that could happen in the middle of the query thread actions.
Lots of comments added in the code describing the reasoning behind the algorithm.

We had to decrease the usage of the query threads, the previous approach was a bit
too optimistic and didn't take into account that a new batch could start before
the previous was completed.