HOPSWORKS-2764
MTR can now be used for upgrade/downgrade tests.
Add a binary ndbmtd_v2 and/or a binary ndb_mgmd_v2 with
either higher or lower version. The MySQL Server will
remain the same version as compiled in the tree.

One also needs to set the environment variable MTR_RONDB_V2
before starting the test.

This also enables schema changes and backups with mixed
versions. To support we define a new table for what versions
are compatible for schema changes and backups