HOPSWORKS-2573
We need to ensure that the LQH_TRANSREQ scan over operation records isn't
negatively affected through the use of operation records in LDM thread
for query thread. It is enough to ensure that tcBlockref is updated
under mutex protection. Added a bit of debugging around this as well.

More debugging support

The changes was focused on handling locked reads, but writes
with multiple replicas have to take into account that the
variable lqhkeyreq_ref only relates to the first node.

Fix for query thread debugging

Handle operations that don't access ACC and TUP, these are used during
node restart when no data has arrived in the fragment yet. They get the
activeCreat set to AC_IGNORED. These should not abort ACC and TUP.

This was properly handled for COMMITs, but not for ABORTs.