HOPSWORKS-2783 Support large transactions in RonDB
Currently large transactions can easily cause job buffer explosion or send buffer
explosion. By introducing batching of abort and commit and complete requests
we ensure that no buffer explosions occur. Also introduced CONTINUEB handling
of releasing records belonging to a large transaction to sustain low latency even
in the context of large transactions.

Previously Take over processing used 1 node operation at a time.
This was safe against overload, but obviously can make large
transactions take ridiculously long time. A transaction with
1 million operations in a 3 replica cluster would have to do
3 million round trips of which 2 million would be to remote
nodes at least. Thus could easily take a couple of hours.

In this patch we instead ensure that we send a batch and when
half of the batch has returned we start sending the next batch.
Using batch sizes of around 1k, we ensure that we can commit
very large transactions at least within a few seconds.

Much of the take over processing also used loops over all
operations, this is changed to handle a batch per real-time
break. Thus ensuring that other transactions are not badly
affected by the large transactions.

To simplify code we don't handle timeouts while we are
still sending batches, similarly while we are transforming
the transaction from the normal transaction handling to the
take over variant.

The take over variant is currently always used when a TC thread
takes over transactions from a failed node. It is also used when
we have a timeout in the commit and complete processing.

We can later introduce other reasons to use this code. It sends
COMMIT and COMPLETE messages to all participants which increases
load on CPUs and networks, but it also decreases latency of
commit operations.

Standardised such that all handling of timeouts used the take over
processing path.

New Feature
............
Introduced a new configuration variable LowLatency that is a boolean
that defaults to off. If not set we get the normal commit behaviour
with linear commit. When set we use normal commit that sends the
commit to all nodes in parallel which should decrease latency at the
expense of higher networking costs, both CPU wise and bandwidth-wise.

This affects the COMMIT phase and COMPLETE phase. It doesn't affect
the prepare phase. Using Low latency means that we will always wait
with releasing the lock until the complete phase.

It is mainly useful in clusters with high latency and clusters with
3-4 replicas.

End New Feature
...............

Fixed problems during development
---------------------------------
Fixed fallthrough issues
Missed initialisations
ndbinfo test case requires a sleep to be safe
Added new host for autotest
Fix conf-Mikael-windows.cnf
Added missing handling of ZSEND_TO_ABORT_HANDLING
Updated configuration files for autotest
Missed initialisation of finish_trans_counter

Missed initialising finish_trans_counter when starting a node failure
takeover transaction. This led to that it never completed unless it
by chance started as 0.

Also missed reset of check_fail_data_process_ongoing after completion
in abort, commit and complete.

Missed incrementing finish_trans_counter when sending ABORTREQ

tcNodedata not properly initialised

In initTcConnectFail we initialise the failData array, but not the
tcNodedata array. execCOMMITCONF and execCOMPLETECONF gets the
replica number by searching for the node id in the tcNodedata array.
Since it isn't initialised, it can contain the node id although the
failData is set to InvalidStatus. This leads to crashes in node
failure handling.

Once started ABORTREQ handling we should focus on failData to report completion

Had to handle all states after node failures in setupFailData

Missed initialisation of copy transaction record when starting toCompleteHandling

Add debug printout in crash situations

When checking and reaching nextTcOperation it is necessary to
reset the check_fail_data_process_ongoing.

Added lots of new checks to ensure that we never allow more than
one thread to work on the transaction to ensure that we don't have
any CONTINUEBs being processed in parallel. This makes it easier
to validate that the code actually works.

Also added a number of comments, only added them in mostly the
abort part, not in the commit and complete part.

Added an optimisation that we check if finish_trans_counter is 0
in the check process to avoid having to loop through a large
transaction when the process of checking is already completed.

Fix of compile and test case issue

Danger with cut and paste

Timeout during release needed to be handled with new state.
No action performed at timeout, simply proceed.
More work on setting send_fail_data_process_ongoing and
check_fail_data_process_ongoing correctly. The handling of
checkFailData_* wasn't correct.

Fix test case with new handling of operations in NDB API.

Prepared for handling timeouts in NDB API. The problem is that with
larger transactions it is fully possible (especially in debug code)
to execute a commit or abort signal for more than 6 minutes. To
handle this we use timer stored on the NdbImpl object. This timer
can be reset by a data node by sending a TC_HBREP signal to the
NDB API as an indication that work is still progressing.

We checked for loop count before we checked if we had already sent
enough signals. This led to various weird issues. But also it meant
that the whole purpose of the patch series got void. So swapped
the order of those in toAbortHandling, toCommitHandling and
toCompleteHandling.

There was a missing return in commit020Lab that could lead to
unnecessary timeouts in Commit processing.

testFK -n CascadeError T1 failed on not releasing all
Signal segments, required an extra sleep before check
since complete processing now takes a bit longer time
to complete and happens in background processing.