HOPSWORKS-2830
Sometimes we get more than 1 COPY_FRAGREQ signal in parallel to DBLQH.
In this case we will often save a corrupt signal since before we save it we have
used the signal object to send another signal. Thus we need to save a copy
of the signal object and use a pointer to the copy rather than the signal object
itself.

Use the same method on COPY_ACTIVEREQ although there is no bug in that
code at the moment.