HOPSWORKS-2783
Added lots of setting of timer and line where it is set.
This will ensure that we won't hit a timeout when still
performing abort/commit work.

Discovered that execCOMMITREQ can have a case where it doesn't
respond, this should not be possible since all operations should
be remaining until we have executed the COMPLETE phase.

Added handling of new state CS_RELEASE in receiving TCKEYREQ.
Should ignore any signals received in this state.

Moved setting of state to CS_ABORTING && abortState = IDLE further
down in releaseAbortResources until we're done with CONTINUEB
handling to avoid allowing new transaction to start while we are
still aborting the previous transaction.