HOPSWORKS-2783
Previously all threads in all nodes worked in parallel
to send LQH_TRANSCONF to the single thread in the master
data node. Now converted local LQH_TRANSREQ to handle
one LDM thread at a time. The sending of LQH_TRANSREQ is
also limited to sending to a maximum of 2 nodes at a
time.

This will limit the risk of overloading the single takeover
thread.