HOPSWORKS-2851
Dropped tables can cause a crash by calling complete_scan_on_table
after the table has been dropped and a new table has been created
using the same i-value of the table.

To avoid this we ensure that all scans remember the schema version
they started the scan against. We check that the schema version is
not a schema version of a new table. It is ok that the schema
version has changed due to an ALTER TABLE.

In the rare case that table version is 0xFFFFFF in the least 24
significant bits (which is the value we set it to after dropping
the table) we keep the values of m_scan_count to avoid crashes
in debug/error build mode.

The schema version cookie is returned in DIH_SCAN_TAB_CONF and
sent to DIH in DIH_SCAN_TAB_COMPLETE_REP. This signal is used
by DBTC, DBSPJ, BACKUP and SUMA.