HOPSWORKS-2886
Query threads don't clean up after scan operations when the TC node
fails. For this to happen it is required that TC and LDMs are on
separate nodes and that the scan operation requires going back to
the API, the failure must also occur while the LDM thread is processing
the scan such that the SCAN_FRAGCONF is sent but not received due to
the node failure. Thus fairly rare, but also important to fix since it
means that the table can't be dropped until the node with the scan is
killed.