HOPSWORKS-2911: Lost signals on ARM
Sending signals between threads on x86 use a technique that is using memory barriers
in a number of places. This technique requires a fair bit of work to get to work on
a new HW platform such as an ARM processor or Power or any other processor. Until this
work have been done we can instead rely on using mutexes for this communication.

The extra overhead this brings costs about 1.5% in performance.