HOPSWORKS-2908
Ensure that GCP stop is enabled by default.
Ensure that DBTC tracks long running transactions to print out
outliers that cause DBTC to block GCPs.

As part of this we decrease the maximum time we can wait for
the application to continue transaction and also set a maximum
timeout for checking deadlock handling. Setting this timeout to
a maximum of 10 seconds.

Added more printouts when GCP stop is close to happening.

Added code to check if DBTC for some reason is making no progress on
handling a GCP. Printouts added to enable better handling of this
issue.