HOPSWORKS-2922: Varsized disk data
Introduce upgrade check if varsized disk data is supported.
Create tables with varsized disk data support if all nodes
in the cluster is supporting it when the table is created.

In principle it could be a decision per node, but it is hard
to know for a node when it is supported and when it isn't.

This means that downgrade will not work very well after creating
the first table with disk data columns.

Make sure that DBTUP is informed about that varsized disk data
rows are to be used. Update the table variables based on the
column types based on this. Also ensure that the proper read
and update routines are used for the disk columns.

Ensure that we reuse as much as possible for read and update routines.