HOPSWORKS-2922: Varsized disk data, step 2
In this patch we add the handling of mapping from size required to
extent bits. In fixed records this was handled by counting free records.
With variable sized rows this needs to use row sizes in bytes instead and
it is important that those limits are independent of the maximum row size
since the maximum can be changed with ALTER TABLE statements.