HOPSWORKS-2922: Varsized disk data, step 6
Handled delete operations where more work was needed to handle
the size of the UNDO log record.

Handled UNDO log execution to handle varsized disk data records
for all operation types.

Fixed a problem with a very long loop giving problems when there are
very many extents in a fragment. This loop could execute millions of
times for a fragment with 1 TB of extents with an extent size of
1 MB.