HOPSWORKS-2937: savepointId destroyed in Commit
In a Commit in DBTUP we start by setting m_commit_disk_callback_page.
This variable is in union with savepointId. savepointId is protected by fragment
mutex and can cause Dirty Reads processed while commit is ongoing to read
the before value even for Dirty Reads part of the same transaction.

Solution is to remove union of those 2 variables.